move xpack rest test back to test
